Update file structure benches; add new `size` function to `Tree`
- There's now a "main" entrypoint into the benching structure
  - the main bench function is located in `benches/bench.rs`
  - it contains the bench-runner too
  - the actual benchers are in the adjacent files
    - they expose public benching functions which the bench-runner runs
